gateware/hsd_ref: fix MGT constrained not applied as out of order
The set_case_analysis constraint in the MGT was
not being applied correctly, because it was parsed
after other constraints that used that generated clock.

As a result, some timing paths were being considered
even though they were static (0 or 1), and timing was
failing badly by >2ns in some paths.