hsd_zcu208: change REFCLK for CLK104 board
Instead of using an SMA pin, which was not possible
due to bank conflicts (MGT must be in the bank as the
SMA, which is not our case), use a general-purpose pin
that is connected to the LPAM connector. This requires
using the CLK_in2 input at the LMK04428B MUX, which is a
software config.