gateware: add BCM AC/DC options
In effect, the only different is the value of a generic
in the system diagram, which is now different
for "bcm_ac" and "bcm_dc".

Note that the "bcm_ac"/"bcm_dc" is not a different target,
but a "variant" which changes only the platform path
for the gateware build.